requirements.txt
Scrapy===1.5.0
Scrapy-Redis==0.6.8
scrapy-fake-useragent==1.1.0
Redis===2.10.6
python-dotenv===0.8.2
oss2===2.4.0
apscheduler===3.5.1
minio===4.0.3
aliyun-mns==1.1.5
bencode.py==2.0.0
pytz
kafka-python